Implemented unit tests

+++ src/Ambev.DeveloperEvaluation.Application/Sales/DeleteSale/DeleteSaleHandler.cs
@@ -1,4 +1,5 @@
-﻿using Ambev.DeveloperEvaluation.Domain.Repositories;
+﻿using Ambev.DeveloperEvaluation.Application.Sales.Common;
+using Ambev.DeveloperEvaluation.Domain.Repositories;
 using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -11,18 +12,22 @@ namespace Ambev.DeveloperEvaluation.Application.Sales.DeleteSale;
 public class DeleteSaleHandler : IRequestHandler<DeleteSaleCommand, DeleteSaleResponse>
 {
     private readonly ISaleRepository _saleRepository;
+    private readonly IMediator _mediator;
     private readonly ILogger<DeleteSaleHandler> _logger;
 
     /// <summary>
     /// Initializes a new instance of DeleteSaleHandler
     /// </summary>
     /// <param name="saleRepository">The sale repository</param>
+    /// <param name="mediator">The Logger instance</param>    
     /// <param name="logger">The Logger instance</param>    
     public DeleteSaleHandler(
         ISaleRepository saleRepository,
+        IMediator mediator,
         ILogger<DeleteSaleHandler> logger)
     {
         _saleRepository = saleRepository;
+        _mediator = mediator;
         _logger = logger;
     }
 
@@ -40,10 +45,17 @@ public class DeleteSaleHandler : IRequestHandler<DeleteSaleCommand, DeleteSaleRe
         if (!validationResult.IsValid)
             throw new ValidationException(validationResult.Errors);
 
-        //TODO configure delete cascade for items
+        var sale = await _saleRepository.GetByIdAsync(request.Id, cancellationToken);
+        if (sale == null)
+            throw new KeyNotFoundException($"Sale with ID {request.Id} not found");
+
+        sale.Cancel(); // raises the domain event
+
         var success = await _saleRepository.DeleteAsync(request.Id, cancellationToken);
         if (!success)
             throw new KeyNotFoundException($"Sale with ID {request.Id} not found");
+
+        await SalesEventsDispatcher.DispatchAsync(sale, _mediator, cancellationToken);
 
         _logger.LogInformation($"Sale with ID {request.Id} deleted successfully");
